Prototype new graph representation
Edge objects are now stored in the allEdges array, only once, no matter if the edge is undirected.
To keep track of what vertices are incident to an edge, the incidenceLists stores indices to the edges
a vertex is incident to. The index of an edge indicates its position in the allEdges array.